rename to .ojo_getters
